[common] fix(color-picker) exposes children prop that defines the picker layout
